+++ messages/__init__.py
@@ -1,3 +1,3 @@
 from .imu_message import IMUMessage
 from .servo_message import ServoMessage
-from .ImageMessage import ImageMessage
+from .image_message import ImageMessage
